feat(shares): Phase 221 FTP 共享详情 API (GET /api/v1/shares/ftp/{id})
- 使用 SqliteShareRepository 真实数据库实现
- JWT 认证，仅 admin 用户可访问
- 验证共享 ID 存在性（404 Not Found）
- 验证协议为 FTP（非 FTP 返回 404）
- 返回字段：id/name/path/description/public/status/created_at/updated_at
- 添加 docs/shares_ftp_get_api.md 文档

+++ src/handlers/shares_ftp_get.rs
@@ -74,7 +74,7 @@ pub async fn get_ftp_share(
     }
 
     // 4. 从数据库查询共享详情
-    match repo.get_share_by_id(share_id as i64) {
+    match repo.get_share_by_id(share_id) {
         Ok(Some(share)) => {
             // 5. 验证协议类型为 FTP
             if share.protocol != "ftp" {
